Add formatting to getPendingTransactions. Move formatTransaction to seperate file
